Corrección de Javadoc
faltaba comentar el método insertarValoracion() de la clase DaoValoración

+++ Symposium_web_git/src/main/java/modelo/ListaArticulo.java
@@ -4,14 +4,15 @@ import java.util.ArrayList;
 
 /**
  * Esta clase permite crear ArrayList de objetos de la clase "Articulo" y trabajar con ellos,
- * introducirlos, mostrarlos por pantalla, borrar un Articulo concreto o vaciar la lista generada.
+ * introducirlos, mostrarlos por pantalla, borrar un Articulo concreto, vaciar la lista generada, etc.
+ * @see Articulo
  * @author Alejandro Moreno
  * @version 1.0
- * @deprecated No se ha llegado a implementar, pero está creada en caso de que en un futuro se desee implementar.
+ * @deprecated No se ha llegado a implementar, pero está creada en caso de que en un futuro se desee hacerlo.
  */
 public class ListaArticulo {
 	
-	private  ArrayList <Articulo> articulos;
+	private ArrayList <Articulo> articulos;
 	
 	/**
 	 * Método constructor de la clase ListaArticulo.
@@ -20,15 +21,14 @@ public class ListaArticulo {
 		articulos=new ArrayList <Articulo>();
 	}
 	/**
-	 * Este método añade objetos "Articulo" a un array "articulos" (y una vez añadido amplia el tamaño de dicho array). 
-	 * @param obra "Articulo" que recibe los parámetros que necesita dicha clase para crear un objeto y 
-	 * 		  los inserta como un objeto de la clase "Articulo" en el array "articulos". 
+	 * Este método añade un objeto <em>Articulo</em> a un array de <em>articulos</em>. 
+	 * @param obra Articulo que se insertará en el array articulos. 
 	 */
 	public void addArticuloLista(Articulo obra) {
 		articulos.add(obra);
 	}
 	/**
-	 * Este método genera una lista con los objetos "Articulo" que se han añadido al array "articulos" 
+	 * Este método genera una lista con los objetos Articulo que se han añadido al array articulos 
 	 * y los muestra por pantalla.
 	 */
 	public void listarArticulos() {
@@ -37,13 +37,13 @@ public class ListaArticulo {
 		}
 	}
 	/**
-	 * Este método permite buscar un libro concreto en el array, comparando el ISBN 
-	 * proporcionado con los ISBN de los objetos "Libro" introucidos en el array "libros".
-	 * @param id Int que que sirve para buscar el Libro concreto en el array comparándolo con
-	 * 		  el ISBN de los objetos "Libro" insertados en la lista.
-	 * @return art Devuelve un objeto "Articulo". 
+	 * Este método permite buscar un articulo concreto en el array, comparando el ISBN 
+	 * proporcionado con los ISBN de los objetos Articulo introucidos en el array articulos.
+	 * @param id long que que sirve para buscar el Articulo concreto en el array comparándolo con
+	 * 		  el ISBN de los objetos de la clase Articulo insertados en la lista.
+	 * @return devuelve un objeto Articulo (si ningún ISBN coincide con los del array, el objeto estará vacío). 
 	 */
-	public Articulo buscarArticulo(int id) {
+	public Articulo buscarArticulo(long id) {
 		Articulo art=new Articulo();
 		boolean cell=false;
 		for (Articulo obj:articulos) {
@@ -66,12 +66,12 @@ public class ListaArticulo {
 	}
 	
 	/**
-	 * Este método elimina un "Articulo" concreto del array "articulos" buscando su ISBN, 
+	 * Este método elimina un Articulo concreto del array articulos buscando su ISBN, 
 	 * manda un mensaje por pantalla tanto si se ha podido eliminar, como si no.
-	 * @param id Int que sirve para buscar el "Articulo" concreto en el array comparándolo con
+	 * @param id long que sirve para buscar el "Articulo" concreto en el array comparándolo con
 	 * 		  el ISBN de los objetos "Articulo" insertados en la lista.
 	 */
-	public void eliminarArticulo(int id) {
+	public void eliminarArticulo(long id) {
 		boolean cell=false;
 		for (Articulo obj:articulos) {
 			if (obj.getISBN()==id) {
