+++ Symposium_web_git/src/main/java/modelo/ListaLibro.java
@@ -3,11 +3,12 @@ package modelo;
 import java.util.ArrayList;
 
 /**
- * Esta clase permite generar ArrayList de objetos "Libro" y trabajar con ellos como un ArrayList, 
- * es decir, insertarlos, mostrarlos por pantalla, borrar un "Libro" concreto o vaciar la lista generada.
+ * Esta clase permite generar ArrayList de objetos Libro y trabajar con ellos como un ArrayList, 
+ * es decir, insertarlos, mostrarlos por pantalla, borrar un Libro concreto, vaciar la lista generada, etc.
+ * @see Libro
  * @author Alejandro Moreno
  * @version 1.0
- * @deprecated No se ha llegado a implementar, pero está creada en caso de que en un futuro se desee implementar.
+ * @deprecated No se ha llegado a implementar, pero está creada en caso de que en un futuro se desee hacerlo.
  */
 public class ListaLibro {
 
@@ -21,16 +22,14 @@ public class ListaLibro {
 		
 	}
 	/**
-	 * Este método añade objetos Libro a un array (y una vez añadido amplia el tamaño de dicho array). 
-	 * @param obra "Libro" que recibe los parámetros que necesita dicha clase para crear un objeto y 
-	 * 		  los inserta como un objeto de la clase "Libro" en el array "libros". 
+	 * Este método añade un Libro nuevo al array,  
+	 * @param obra Libro que se insertará en el array libros. 
 	 */
 	public void addLibroLista(Libro obra) {
 		libros.add(obra);
 	}
 	/**
-	 * Este método genera una lista con los objetos "Libro" que se han añadido al array "libros" 
-	 * y los muestra por pantalla.
+	 * Este método genera una lista con los objetos Libro que se han añadido al array y los muestra por pantalla.
 	 */
 	public void listarLibros() {
 		for(Libro obj:libros) {
@@ -39,12 +38,11 @@ public class ListaLibro {
 	}
 	/**
 	 * Este método permite buscar un libro concreto en el array, comparando el ISBN 
-	 * proporcionado con los ISBN de los objetos "Libro" introucidos en el array "libros".
-	 * @param id Int que que sirve para buscar el Libro concreto en el array comparándolo con
-	 * 		  el ISBN de los objetos "Libro" insertados en la lista.
-	 * @return lib Devuelve el objeto "Libro" si lo encuentra en el array. 
+	 * proporcionado con los ISBN de los libros introucidos en el array.
+	 * @param id long que recoge la ISBN mediante la cual se buscará el Libro concreto en el array.
+	 * @return devuelve un objeto Libro si lo encuentra en el array (si no lo encuentra devolverá un objeto con valores vacíos). 
 	 */
-	public Libro buscarLibro(int id) {
+	public Libro buscarLibro(long id) {
 		Libro lib=new Libro();
 		boolean cell=false;
 		for (Libro obj:libros) {
@@ -59,18 +57,16 @@ public class ListaLibro {
 		return lib;
 	}
 	/**
-	 * Este método elimina completamente los objetos "Libro" que hay dentro del array "libros"
+	 * Este método elimina completamente los objetos que hay dentro del array.
 	 */
 	public void borrarListaCompleta() {
 		libros.clear();
 	}
 	/**
-	 * Este método elimina un "Libro" concreto del array "libros" buscando su ISBN, 
-	 * manda un mensaje por pantalla tanto si se ha podido eliminar, como si no.
-	 * @param id Int que sirve para buscar el Libro concreto en el array comparándolo con
-	 * 		  el ISBN de los objetos "Libro" insertados en la lista.
+	 * Este método elimina un Libro concreto del array buscando su ISBN, manda un mensaje por pantalla tanto si se ha podido eliminar, como si no.
+	 * @param id long que recoge la ISBN que se buscará en el array.
 	 */
-	public void eliminarLibro(int id) {
+	public void eliminarLibro(long id) {
 		boolean cell=false;
 		for (Libro obj:libros) {
 			if (obj.getISBN()==id) {
